% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BallMapper.R
\name{colorByAllVariables}
\alias{colorByAllVariables}
\title{Produce a collection of png files with mapper graphs colored by following coordinates (so that the number of files is the same as the number of coordinates).}
\usage{
colorByAllVariables(outputFromBallMapper, points,
  fileNamePrefix = "output_", defaultXResolution = 512,
  defaultYResolution = 512)
}
\arguments{
\item{outputFromBallMapper}{an output from the BallMapper function}

\item{points, }{a collection of input points in a form of a data frame used to create Ball Mapper graph.}

\item{fileNamePrefix}{a prefix of a file name. A plot that uses i-th variable as a coloring will contain this string as a prefix followed by the number i. Set to "output_" by default.}

\item{defaultXResolution}{store a default resolution of image in x direction. Set to 512 by default.}

\item{defaultYResolution}{store a default resolution of image in y direction. Set to 512 by default.}
}
\value{
none.
}
\description{
Produce a collection of png files with mapper graphs colored by following coordinates (so that the number of files is the same as the number of coordinates).
}
