% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BallMapper.R
\name{colorByStDevValueOfOtherVariable}
\alias{colorByStDevValueOfOtherVariable}
\title{Produce a new coloring vector being a standard deviation of values of given
function at points covererd by each vertex of Ball Mapper graph.}
\usage{
colorByStDevValueOfOtherVariable(outputFromBallMapper, newFunctionOnPoints)
}
\arguments{
\item{outputFromBallMapper}{an output from the BallMapper function}

\item{newFunctionOnPoints}{values of function on points.}
}
\value{
Vector of function values on vertices on Ball Mapper graph.
var <- seq(from=0,to=6.3,by=0.1)
points <- as.data.frame( cbind( sin(var),cos(var) ) )
values <- as.data.frame(sin(var))
l <- BallMapper(points, values, 0.25)
ColorIgraphPlot(l)
new_coloring <- colorByStDevValueOfOtherVariable(l,sin(var))
l$coloring <- new_coloring
ColorIgraphPlot(l)
}
\description{
Produce a new coloring vector being a standard deviation of values of given
function at points covererd by each vertex of Ball Mapper graph.
}
