% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BallMapper.R
\name{coordinates_of_points_in_subcollection}
\alias{coordinates_of_points_in_subcollection}
\title{This is an auxiliery function. It take the coordinates of points, ids of subset of points, and number of coordinate, and return a sorted vector of the given coodrinate in the considered points.
For instance, given the collection of points:
1 2 3
4 5 6
7 8 9
and which_subset = 2,3
and number_of_coordinate = 2
the procedure below will return the vector [2,5,8].}
\usage{
coordinates_of_points_in_subcollection(points, which_subset,
  number_of_coordinate)
}
\arguments{
\item{points}{is a collection of input points in a form of a data frame. The same one as on the input of the Ball Mapper.}

\item{which_subset}{Indices of points in the given subset.}

\item{number_of_coordinate}{which coordinate of the consired points to export.}
}
\value{
the sorted vector of values of a given variable at the collection of points.
var <- seq(from=0,to=6.3,by=0.1)
points <- as.data.frame( cbind( sin(var),cos(var) ) )
values <- as.data.frame(sin(var))
l <- BallMapper(points, values, 0.25)
coordinates_of_points_in_subcollection(points,c(6,7,8),1)
}
\description{
This is an auxiliery function. It take the coordinates of points, ids of subset of points, and number of coordinate, and return a sorted vector of the given coodrinate in the considered points.
For instance, given the collection of points:
1 2 3
4 5 6
7 8 9
and which_subset = 2,3
and number_of_coordinate = 2
the procedure below will return the vector [2,5,8].
}
