% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BallMapper.R
\name{find_dominant_difference_using_averages}
\alias{find_dominant_difference_using_averages}
\title{This procedure take two subset of points (that come from the vertices of Ball Mapper) and return
the coordinates on which the averages of those two collections differs most. To ballance the effect
of potentially different orders of magnitude of data in column, we divide the difference in means by the mean of the whole column.}
\usage{
find_dominant_difference_using_averages(points, subset1, subset2)
}
\arguments{
\item{points}{a collection of input points in a form of a data frame. The same one as on the input of the Ball Mapper.}

\item{subset1}{First subset of ids of points.}

\item{subset2}{Second subset of ids of points.}
}
\value{
Vector of corrdinate ids with the absolute value of difference between averages, ordered according to the second variable.
var <- seq(from=0,to=6.3,by=0.1)
points <- as.data.frame( cbind( sin(var),cos(var) ) )
values <- as.data.frame(sin(var))
l <- BallMapper(points, values, 0.25)
g1 <- c(1,21
g2 <- c(11,12)
find_dominant_difference_using_averages(points,g1,g2)
}
\description{
This procedure take two subset of points (that come from the vertices of Ball Mapper) and return
the coordinates on which the averages of those two collections differs most. To ballance the effect
of potentially different orders of magnitude of data in column, we divide the difference in means by the mean of the whole column.
}
