% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BallMapper.R
\name{normalize_to_average_0_stdev_1}
\alias{normalize_to_average_0_stdev_1}
\title{This function normalize each column (variable) of the input dataset so that the the average of the normalized column is 0 and its standard deviation is 1.}
\usage{
normalize_to_average_0_stdev_1(points)
}
\arguments{
\item{points, }{a collection of input points in a form of a data frame.}
}
\value{
Nowmalized collectpion of points.
}
\description{
This function normalize each column (variable) of the input dataset so that the the average of the normalized column is 0 and its standard deviation is 1.
}
\examples{
var <- seq(from=0,to=6.3,by=0.1)
points <- as.data.frame( cbind( sin(var),cos(var) ) )
normalized_points <- normalize_to_average_0_stdev_1 (points)
}
