% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BallMapper.R
\name{normalize_to_min_0_max_1}
\alias{normalize_to_min_0_max_1}
\title{This function normalize each column (variable) of the input dataset so that the maximum is mapped to one, minimum to zero, and the intermediate values linearly to the appropriate points in the interval (0,1).}
\usage{
normalize_to_min_0_max_1(points)
}
\arguments{
\item{points, }{a collection of input points in a form of a data frame.}
}
\value{
Normalized collection of points.
}
\description{
This function normalize each column (variable) of the input dataset so that the maximum is mapped to one, minimum to zero, and the intermediate values linearly to the appropriate points in the interval (0,1).
}
\examples{
var <- seq(from=0,to=6.3,by=0.1)
points <- as.data.frame( cbind( sin(var),cos(var) ) )
normalized_points <- normalize_to_min_0_max_1 (points)
}
