% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BallMapper.R
\name{points_covered_by_landmarks}
\alias{points_covered_by_landmarks}
\title{This function returns a list of points covered by the given collection of landmarks.}
\usage{
points_covered_by_landmarks(outputFromBallMapper, numbers_of_landmarks)
}
\arguments{
\item{outputFromBallMapper}{an output from the BallMapper function}

\item{numbers_of_landmarks}{a vector containnig the numbers of landmarks under consideration.}
}
\value{
A vector of points covered by the landmarks given in numbers_of_landmarks.
}
\description{
This function returns a list of points covered by the given collection of landmarks.
}
