% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BallMapper.R
\name{readBallMapperGraphFromFile}
\alias{readBallMapperGraphFromFile}
\title{This procedure read the BallMapper object from file. The parameter of the file
is filename. We assume that files:
filename_vertices
filename_edges
filename_edges_strength
filename_points_covered_by_landmarks
filename_landmarks
filename_coloring}
\usage{
readBallMapperGraphFromFile(filename)
}
\arguments{
\item{filename}{prefix of the name of the file containing elements of Ball Mapper graph.}
}
\value{
BallMapper object
var <- seq(from=0,to=6.3,by=0.1)
points <- as.data.frame( cbind( sin(var),cos(var) ) )
values <- as.data.frame(sin(var))
l <- BallMapper(points, values, 0.25)
storeBallMapperGraphInFile(l,"my_favorite_BM_graph")
l_prime <- readBallMapperGraphFromFile("my_favorite_BM_graph")
}
\description{
This procedure read the BallMapper object from file. The parameter of the file
is filename. We assume that files:
filename_vertices
filename_edges
filename_edges_strength
filename_points_covered_by_landmarks
filename_landmarks
filename_coloring
}
