% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BallMapper.R
\name{simpleDynamicNetwork}
\alias{simpleDynamicNetwork}
\title{This is a simple example of dynamic visualization using networkD3 library.
This version do not implement coloring of vertices, just give a general overview of the edges.}
\usage{
simpleDynamicNetwork(outputFromBallMapper, storeAsHtml = FALSE)
}
\arguments{
\item{outputFromBallMapper, }{an output from BallMapper function.}

\item{storeAsHtml, }{if set true, it will store the graph in HTML file.}
}
\value{
None
}
\description{
This is a simple example of dynamic visualization using networkD3 library.
This version do not implement coloring of vertices, just give a general overview of the edges.
}
\examples{
var <- seq(from=0,to=6.3,by=0.1)
points <- as.data.frame( cbind( sin(var),cos(var) ) )
values <- as.data.frame( sin(var) )
epsilon <- 0.25
l <- BallMapper(points,values,epsilon)
simpleDynamicNetwork(l)
}
