% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BallMapper.R
\name{storeBallMapperGraphInFile}
\alias{storeBallMapperGraphInFile}
\title{This procedure store the Ball Mapper graph in a file in the following format:}
\usage{
storeBallMapperGraphInFile(outputFromBallMapper, filename = "BM_graph")
}
\arguments{
\item{outputFromBallMapper}{output from the BallMapper procerure.}

\item{filename}{the name of the file to store the data.}
}
\value{
None
var <- seq(from=0,to=6.3,by=0.1)
points <- as.data.frame( cbind( sin(var),cos(var) ) )
values <- as.data.frame(sin(var))
l <- BallMapper(points, values, 0.25)
storeBallMapperGraphInFile(l,"my_favorite_BM_graph")
}
\description{
This procedure store the Ball Mapper graph in a file in the following format:
}
